requirements.txt
Flask==2.2.3
peewee==3.17.8
werkzeug==2.2.3
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-dotenv==1.0.1
pymongo==4.11.3
mongoengine==0.29.1
certifi==2025.1.31
twilio==9.5.2
APScheduler==3.11.0